+++ src/Commands/Tags/Tag.ts
@@ -2,7 +2,7 @@ import { Command } from "../../Structures/Command";
 import { Message } from "discord.js";
 import { pool } from "../../Structures/Database/Mongo";
 import { Tags } from "../../lib/types/Collections";
-import KhafraClient from "../../Bot/KhafraBot";
+import { KhafraClient } from "../../Bot/KhafraBot";
 
 export default class extends Command {
     constructor() {